refactor: extract useDateTimeInput hook, dedupe 3 modals' parse-as-you-type logic
Three modals reimplemented identical natural-language date+time entry
+ live preview logic verbatim:
  • EditCheckModal
  • EditEventModal
  • CreateModal (the "Interest Check" mode)

Each had its own copy of:
  const parsedDate = whenInput ? parseNaturalDate(whenInput) : null;
  const parsedTime = whenInput ? parseNaturalTime(whenInput) : null;
  const whenPreview = parsedDate || parsedTime ? ... : null;

Now in src/shared/hooks/useDateTimeInput.ts. Each modal pulls the
state + parsed values + preview string from the hook directly.

Net: -27 LOC across 3 modals, single source of truth for parse logic.
e2e still 19/19.
